# 02_PWA-Studio-Pro/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PWA Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: string;
  name: string;
  description: string;
  thumbnailUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectSummary {
  id: string;
  name: string;
  description: string;
  thumbnailUrl?: string | null;
  layoutCount: number;
  widgetCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface LayoutSummary {
  id: string;
  projectId: string;
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectDetail {
  id: string;
  name: string;
  description: string;
  thumbnailUrl?: string | null;
  layouts: LayoutSummary[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  description?: string;
  thumbnailUrl?: string | null;
}

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  thumbnailUrl?: string | null;
}

/**
 * react-grid-layout state (array of layout items)
 */
export type LayoutGridLayout = { [key: string]: unknown };

/**
 * React Flow state (nodes and edges)
 */
export type LayoutFlowGraph = { [key: string]: unknown };

export interface Layout {
  id: string;
  projectId: string;
  name: string;
  /** react-grid-layout state (array of layout items) */
  gridLayout: LayoutGridLayout;
  /** React Flow state (nodes and edges) */
  flowGraph: LayoutFlowGraph;
  createdAt: string;
  updatedAt: string;
}

export type CreateLayoutBodyGridLayout = { [key: string]: unknown };

export type CreateLayoutBodyFlowGraph = { [key: string]: unknown };

export interface CreateLayoutBody {
  name: string;
  gridLayout?: CreateLayoutBodyGridLayout;
  flowGraph?: CreateLayoutBodyFlowGraph;
}

export type UpdateLayoutBodyGridLayout = { [key: string]: unknown };

export type UpdateLayoutBodyFlowGraph = { [key: string]: unknown };

export interface UpdateLayoutBody {
  name?: string;
  gridLayout?: UpdateLayoutBodyGridLayout;
  flowGraph?: UpdateLayoutBodyFlowGraph;
}

/**
 * Zod-compatible JSON schema for widget config
 */
export type WidgetRegistryEntrySchema = { [key: string]: unknown };

export interface WidgetRegistryEntry {
  id: string;
  slug: string;
  name: string;
  version: string;
  description: string;
  /** Zod-compatible JSON schema for widget config */
  schema: WidgetRegistryEntrySchema;
  tags: string[];
  isBuiltin: string;
}

export type RegisterWidgetBodySchema = { [key: string]: unknown };

export interface RegisterWidgetBody {
  slug: string;
  name: string;
  version?: string;
  description?: string;
  schema?: RegisterWidgetBodySchema;
  tags?: string[];
  isBuiltin?: string;
}

export type ConversationMessageRole = typeof ConversationMessageRole[keyof typeof ConversationMessageRole];


export const ConversationMessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface ConversationMessage {
  id: string;
  projectId: string;
  role: ConversationMessageRole;
  content: string;
  model: string;
  tokensUsed?: number | null;
  createdAt: string;
}

export type AppendConversationBodyRole = typeof AppendConversationBodyRole[keyof typeof AppendConversationBodyRole];


export const AppendConversationBodyRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface AppendConversationBody {
  role: AppendConversationBodyRole;
  content: string;
  model?: string;
  tokensUsed?: number | null;
}

export type DetectedFeatureConfidence = typeof DetectedFeatureConfidence[keyof typeof DetectedFeatureConfidence];


export const DetectedFeatureConfidence = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface DetectedFeature {
  id: string;
  suggestedSlug: string;
  suggestedName: string;
  description: string;
  sourceFiles: string[];
  confidence: DetectedFeatureConfidence;
  tags: string[];
}

export interface PwaAnalysisResult {
  appName: string;
  appDescription: string;
  framework: string;
  themeColor: string;
  totalFiles: number;
  features: DetectedFeature[];
}

export type PwaExtractFeatureSchema = { [key: string]: unknown };

export interface PwaExtractFeature {
  slug: string;
  name: string;
  description?: string;
  tags?: string[];
  schema?: PwaExtractFeatureSchema;
}

export interface PwaExtractBody {
  projectName: string;
  projectDescription?: string;
  features: PwaExtractFeature[];
}

export interface PwaExtractResult {
  projectId: string;
  layoutId: string;
  widgetsCreated: number;
  widgetsSkipped: number;
}

export interface ChatMessageBody {
  /** @minLength 1 */
  content: string;
}

export interface WidgetTypeStat {
  widgetType: string;
  count: number;
}

export interface DashboardSummary {
  projectCount: number;
  widgetCount: number;
  layoutCount: number;
  conversationCount: number;
  recentProjects: ProjectSummary[];
  topWidgetTypes: WidgetTypeStat[];
}

export interface ProjectActivity {
  projectId: string;
  recentLayouts: LayoutSummary[];
  recentMessages: ConversationMessage[];
}

export type UserSettingsTheme = typeof UserSettingsTheme[keyof typeof UserSettingsTheme];


export const UserSettingsTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export type UserSettingsApiOverrides = { [key: string]: unknown };

export interface UserSettings {
  id: number;
  theme: UserSettingsTheme;
  activeModel: string;
  apiOverrides: UserSettingsApiOverrides;
  updatedAt: string;
}

export type UpdateSettingsBodyTheme = typeof UpdateSettingsBodyTheme[keyof typeof UpdateSettingsBodyTheme];


export const UpdateSettingsBodyTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export type UpdateSettingsBodyApiOverrides = { [key: string]: unknown };

export interface UpdateSettingsBody {
  theme?: UpdateSettingsBodyTheme;
  activeModel?: string;
  apiOverrides?: UpdateSettingsBodyApiOverrides;
}

export type ListConversationsParams = {
limit?: number;
offset?: number;
};

export type AnalyzePwaZipBody = {
  file: Blob;
};