# 03_PWA-Studio-mobile/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: APP Studio API — AI-Assisted Multi-Agent PWA Creator
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, all]
            default: all
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Canvas ──────────────────────────────────────────────────────────────
...
  /projects/{projectId}/activity:
    get:
      operationId: getActivityFeed
      tags: [agents]
      summary: Get recent agent activity feed for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity feed events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
    post:
      operationId: createActivity
      tags: [agents]
      summary: Log an activity event
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityEventInput"
      responses:
        "201":
          description: Created activity event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityEvent"

  # ── Settings ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, building, testing, deployed, archived]
        framework:
          type: ["string", "null"]
        githubRepo:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        taskCount:
          type: integer
        completedTaskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        framework:
          type: string
        githubRepo:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [active, building, testing, deployed, archived]
        framework:
          type: string
        githubRepo:
          type: string
        previewUrl:
          type: string
...
    ActivityEvent:
      type: object
      required: [id, projectId, agentRole, type, message, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        agentRole:
          type: string
          enum: [director, design, image, builder, tester, deployer, reviewer, system]
        type:
          type: string
          enum: [info, success, warning, error, progress]
        message:
          type: string
        detail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ActivityEventInput:
      type: object
      required: [agentRole, type, message]
      properties:
        agentRole:
          type: string
          enum: [director, design, image, builder, tester, deployer, reviewer, system]
        type:
          type: string
          enum: [info, success, warning, error, progress]
        message:
          type: string
        detail:
          type: string
...